float2string #3 (lyxrc)
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@9432 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/frontends/gnome/lyx_gui.C
@@ -72,11 +72,11 @@ namespace {
 bool finished = false;
 
 /// estimate DPI from X server
-float getDPI()
+int getDPI()
 {
 	Screen * scr = ScreenOfDisplay(fl_get_display(), fl_screen);
-	return ((HeightOfScreen(scr) * 25.4 / HeightMMOfScreen(scr)) +
-		(WidthOfScreen(scr) * 25.4 / WidthMMOfScreen(scr))) / 2;
+	return int(((HeightOfScreen(scr) * 25.4 / HeightMMOfScreen(scr)) +
+		(WidthOfScreen(scr) * 25.4 / WidthMMOfScreen(scr))) / 2);
 }
